+++ projects/ngx-opendesign/src/public-api.ts
@@ -48,6 +48,8 @@ export * from './lib/pagination/pagination';
 export * from './lib/file-input/file-input';
 export * from './lib/image-preview/image-preview';
 export * from './lib/image-slider/image-slider';
+export * from './lib/carousel/carousel';
+export * from './lib/gallery/gallery';
 export * from './lib/divider/divider';
 export * from './lib/fab/fab';
 export * from './lib/datepicker/datepicker';